refine datafile display

+++ app/controllers/datafiles_controller.rb
@@ -14,12 +14,16 @@ class DatafilesController < ApplicationController
   # GET /datafiles
   # GET /datafiles.json
   def index
-    @dataset = Dataset.find_by_key(params[:dataset_id])
-    @datafiles = Datafile.all
-    @datafiles.each do |datafile|
-      datafile.destroy unless ( (datafile.binary && datafile.binary.file) || (datafile.medusa_path && datafile.medusa_path != "") )
+
+    if params.has_key?(:dataset_id)
+      @dataset = Dataset.find_by_key(params[:dataset_id])
+      @datafiles = Datafile.all
+      @datafiles.each do |datafile|
+        datafile.destroy unless ( (datafile.binary && datafile.binary.file) || (datafile.medusa_path && datafile.medusa_path != "") )
+      end
+      authorize! :edit, @dataset
     end
-    authorize! :edit, @dataset
+
   end
 
   # GET /datafiles/1
